Style: 한 줄 소개 css 수정
